SPA-41 Add unit tests for if statements and fix unsigned int bug

+++ Team03/Code03/src/spa/src/sp/extractor/abstraction/follows_abstraction_extractor.cpp
@@ -51,7 +51,7 @@ void FollowsAbstractionExtractor::RetrieveFromChildren(
     std::vector<FollowsAbstraction>* follows_abstractions,
     const std::vector<std::shared_ptr<TNode>>& children,
     std::unordered_map<TNode, StatementEntity*>& t_node_stmt_ent_umap) const {
-  for (int i = 0; i < children.size() - 1; i++) {
+  for (int i = 0; i < (int)children.size() - 1; i++) {
     auto lhs = t_node_stmt_ent_umap[*children[i]];
     auto rhs = t_node_stmt_ent_umap[*children[i + 1]];
     follows_abstractions->emplace_back(*lhs, *rhs);
